pyproject.toml
[project]
name = "options-backtester"
version = "0.2.0"
description = "Modular backtesting engine for options strategies"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=0.10.0",
    "pyarrow>=15.0.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "plotly>=5.18.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.3.0"]

[tool.ruff]
line-length = 100
target-version = "py311"
